+++ projeto/backend/controllers/VendaController.js
@@ -8,30 +8,34 @@ exports.get = async (req, res) => {
   if (AutorizacaoService.validarRoles(req, [ROLES.VENDEDOR, ROLES.CLIENTE, ROLES.ADMIN])) {
     let id = req.params.id;
     let erro = true;
-    try {
-      let registro = await VendaService.getVendabyId(id);
-      // Se for apenas cliente só pode recuperar a própria compra
-      if ((registro) && (!AutorizacaoService.validarRoles(req, [ROLES.VENDEDOR, ROLES.ADMIN]))){
-        let email = AutorizacaoService.getEmail(req);
-        let cliente = await ClienteService.findOne({email: email});
-        if (cliente) {
-          //https://stackoverflow.com/questions/11637353/comparing-mongoose-id-and-strings
-          erro = !cliente._id.equals(registro.id_cliente);
-        }  
-      } else {
-        erro = false;
-      }
-      if (erro) {
-        res.status(403).json({ error: 'Acesso negado' });
-      } else {
-        if (registro) {
-          res.json(registro);
+    if (id.length == 24) {
+      try {
+        let registro = await VendaService.getVendabyId(id);
+        // Se for apenas cliente só pode recuperar a própria compra
+        if ((registro) && (!AutorizacaoService.validarRoles(req, [ROLES.VENDEDOR, ROLES.ADMIN]))) {
+          let email = AutorizacaoService.getEmail(req);
+          let cliente = await ClienteService.findOne({ email: email });
+          if (cliente) {
+            //https://stackoverflow.com/questions/11637353/comparing-mongoose-id-and-strings
+            erro = !cliente._id.equals(registro.id_cliente);
+          }
         } else {
-          res.status(404).json({});        
-        }  
+          erro = false;
+        }
+        if (erro) {
+          res.status(403).json({ error: 'Acesso negado' });
+        } else {
+          if (registro) {
+            res.json(registro);
+          } else {
+            res.status(404).json({});
+          }
+        }
+      } catch (error) {
+        res.status(500).json({ error: error.message });
       }
-    } catch (error) {
-      res.status(500).json({ error: error.message });
+    } else {
+      res.status(400).json({});
     }
   } else {
     res.status(403).json({ error: 'Acesso negado' });
@@ -47,19 +51,19 @@ exports.getAll = async (req, res) => {
     try {
       //Se está fazendo filtro por cliente só pode recuperar os registros do cliente que é o usuário logado
       if (filtroCliente) {
-         let email = AutorizacaoService.getEmail(req);
-         let cliente = await ClienteService.findOne({email: email});
-         if (cliente) {
-           id_cliente = cliente._id;
-         } else {
-           erro = true;
-         }
+        let email = AutorizacaoService.getEmail(req);
+        let cliente = await ClienteService.findOne({ email: email });
+        if (cliente) {
+          id_cliente = cliente._id;
+        } else {
+          erro = true;
+        }
       }
       let registros = [];
-      if (!erro){
-         //https://stackoverflow.com/questions/6912584/how-to-get-get-query-string-variables-in-express-js-on-node-js
-         registros = await VendaService.getAllVendas(id_cliente, req.query.ano, req.query.agrupar);
-      }  
+      if (!erro) {
+        //https://stackoverflow.com/questions/6912584/how-to-get-get-query-string-variables-in-express-js-on-node-js
+        registros = await VendaService.getAllVendas(id_cliente, req.query.ano, req.query.agrupar);
+      }
 
       res.json(registros);
     } catch (err) {
@@ -87,15 +91,19 @@ exports.delete = async (req, res) => {
   if (AutorizacaoService.validarRoles(req, [ROLES.VENDEDOR, ROLES.ADMIN])) {
     let id = req.params.id;
 
-    try {
-      const registro = await VendaService.deleteVenda(id);
-      if (registro) {
-        res.json(registro);
-      } else {
-        res.status(404).json({});        
-      }  
-    } catch (error) {
-      res.status(500).json({ error: error.message });
+    if (id.length == 24) {
+      try {
+        const registro = await VendaService.deleteVenda(id);
+        if (registro) {
+          res.json(registro);
+        } else {
+          res.status(404).json({});
+        }
+      } catch (error) {
+        res.status(500).json({ error: error.message });
+      }
+    } else {
+      res.status(400).json({});
     }
   } else {
     res.status(403).json({ error: 'Acesso negado' });
@@ -111,8 +119,8 @@ exports.getRelatorioListagem = async (req, res) => {
       await PDFService.gerarPDF(res, 'Vendas', [
         { label: 'Data', property: 'data', width: 70, renderer: null },
         { label: 'Vendedor', property: 'vendedor', width: 200, renderer: null },
-        { label: 'Valor', property: 'valor', width: 70, renderer: (value) => {return RelatorioUtilService.getDinheiro(value)} },
-        { label: 'Lucro', property: 'lucro', width: 70, renderer: (value) => {return RelatorioUtilService.getDinheiro(value, true)} },], dados);
+        { label: 'Valor', property: 'valor', width: 70, renderer: (value) => { return RelatorioUtilService.getDinheiro(value) } },
+        { label: 'Lucro', property: 'lucro', width: 70, renderer: (value) => { return RelatorioUtilService.getDinheiro(value, true) } },], dados);
     } catch (err) {
       return res.status(500).json({ error: err.message });
     }
@@ -129,7 +137,7 @@ exports.getExcelListagem = async (req, res) => {
       res.set({
         'Content-Disposition': 'attachment; filename=Vendas.xlsx',
         'Content-Type': "application/vnd.openxmlformats-officedocument.spreadsheetml.sheet"
-      });          
+      });
       res.xls('Vendas.xlsx', registros);
     } catch (err) {
       return res.status(500).json({ error: err.message });
